Chore: refactored the approach by shifting the drag logic in inserter-dragable-blocks and added the onDrag and onClick close support to media and patterns
