+++ packages/block-editor/src/components/inserter-list-item/index.js
@@ -15,8 +15,6 @@ import {
 } from '@wordpress/blocks';
 import { __experimentalTruncate as Truncate } from '@wordpress/components';
 import { ENTER, isAppleOS } from '@wordpress/keycodes';
-import { useSelect } from '@wordpress/data';
-import { useViewportMatch } from '@wordpress/compose';
 
 /**
  * Internal dependencies
@@ -24,8 +22,6 @@ import { useViewportMatch } from '@wordpress/compose';
 import BlockIcon from '../block-icon';
 import { InserterListboxItem } from '../inserter-listbox';
 import InserterDraggableBlocks from '../inserter-draggable-blocks';
-import { store as blockEditorStore } from '../../store';
-import { unlock } from '../../lock-unlock';
 
 function InserterListItem( {
 	className,
@@ -58,17 +54,6 @@ function InserterListItem( {
 		( isReusableBlock( item ) && item.syncStatus !== 'unsynced' ) ||
 		isTemplatePart( item );
 
-	const { setInserterIsOpened } = useSelect( ( select ) => {
-		const { getSettings } = unlock( select( blockEditorStore ) );
-
-		return {
-			setInserterIsOpened:
-				getSettings().__experimentalSetIsInserterOpened,
-		};
-	}, [] );
-
-	const isMobileViewport = useViewportMatch( 'medium', '<' );
-
 	return (
 		<InserterDraggableBlocks
 			isEnabled={ isDraggable && ! item.isDisabled }
@@ -89,9 +74,6 @@ function InserterListItem( {
 						if ( onDragStart ) {
 							onHover( null );
 							onDragStart( event );
-						}
-						if ( isMobileViewport ) {
-							setInserterIsOpened( false );
 						}
 					} }
 					onDragEnd={ ( event ) => {
